moves stateful components into /views and stateless components stay in /components
