+++ src/components/NotFoundPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export const NotFoundPage: React.SFC = () => {
+export const NotFoundPage: React.FC = () => {
     return <p>Page not found.</p>;
 };
