+++ src/components/Room.tsx
@@ -4,7 +4,7 @@ import Constants from '../constants/Constants';
 import * as Types from '../constants/Types';
 import socket from '../constants/socket-context';
 import Messaging from './Messaging';
-import FileTransfer from './FileTransfer';
+import FileTransfer from '../views/FileTransfer';
 
 type RoomProps = {
     currentRoom: Types.Room;
